% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{can_run}
\alias{can_run}
\title{Check if we can run a given operation node}
\usage{
can_run(op_node)
}
\arguments{
\item{dist_node}{a greta operation node}
}
\value{
a logical vector of length 1
}
\description{
Check if we can run a given operation node
}
