% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{can_sample}
\alias{can_sample}
\title{Check if we can sample from a given distribution node}
\usage{
can_sample(dist_node)
}
\arguments{
\item{dist_node}{a greta distribution node}
}
\value{
a logical vector of length 1
}
\description{
Check if we can sample from a given distribution node
}
