% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{dag_status}
\alias{dag_status}
\title{Get the status of the dag}
\usage{
dag_status(dag)
}
\arguments{
\item{dag}{the DAG of the greta model}
}
\value{
A named list
}
\description{
Get the status of the dag
}
