% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_name}
\alias{dist_name}
\title{Get the name of distribution the distibution node follows}
\usage{
dist_name(dist_node)
}
\arguments{
\item{dist_node}{a greta distibution node}
}
\value{
the name of dist
}
\description{
Get the name of distribution the distibution node follows
}
