% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distribution_nodes}
\alias{distribution_nodes}
\title{Get the distibution nodes of a greta dag}
\usage{
distribution_nodes(dag)
}
\arguments{
\item{dag}{a greta dag}
}
\value{
a list of greta nodes
}
\description{
Get the distibution nodes of a greta dag
}
