% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index}
\alias{index}
\title{What index are nodes in the dag}
\usage{
index(node, dag)
}
\arguments{
\item{node}{a greta node}

\item{dag}{a greta dag}
}
\value{
a numeric vector
}
\description{
What index are nodes in the dag
}
