% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_functions.R
\name{insert_parameters}
\alias{insert_parameters}
\title{Insert values of parameters into a DAG}
\usage{
insert_parameters(parameters, dag)
}
\arguments{
\item{parameters}{a names list of parameters and their values}

\item{dag}{a the parameters are going to be inserted into}
}
\value{
the altered DAG
}
\description{
Insert values of parameters into a DAG
}
