% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.data_node}
\alias{is.data_node}
\title{is an object a greta data node}
\usage{
is.data_node(x)
}
\arguments{
\item{x}{an object}
}
\value{
a length 1 logical vector
}
\description{
is an object a greta data node
}
