% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.distribution_node}
\alias{is.distribution_node}
\title{is an object a greta distibution node}
\usage{
is.distribution_node(x)
}
\arguments{
\item{x}{an object}
}
\value{
a length 1 logical vector
}
\description{
is an object a greta distibution node
}
