% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.greta_model}
\alias{is.greta_model}
\title{is an object a greta model}
\usage{
is.greta_model(x)
}
\arguments{
\item{x}{an object}
}
\value{
a length 1 logical vector
}
\description{
is an object a greta model
}
