% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.operation_node}
\alias{is.operation_node}
\title{is an object a greta operation node}
\usage{
is.operation_node(x)
}
\arguments{
\item{x}{an object}
}
\value{
a length 1 logical vector
}
\description{
is an object a greta operation node
}
