% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep}
\alias{keep}
\title{keep the elements of the list which pass a predicate}
\usage{
keep(list, pred)
}
\arguments{
\item{list}{a list}

\item{pred}{a predicate: a function that takes an element of the list
and outputs a length 1 logical vector}
}
\value{
the reduced list
}
\description{
keep the elements of the list which pass a predicate
}
