% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{node_names}
\alias{node_names}
\title{Get the names of the nodes in a dag}
\usage{
node_names(dag)
}
\arguments{
\item{dag}{a greta dag}
}
\value{
the names of the nodes
}
\description{
Get the names of the nodes in a dag
}
