% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{operation_nodes}
\alias{operation_nodes}
\title{Get the operation nodes of a greta dag}
\usage{
operation_nodes(dag)
}
\arguments{
\item{dag}{a greta dag}
}
\value{
a list of greta nodes
}
\description{
Get the operation nodes of a greta dag
}
