% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{run_add}
\alias{run_add}
\title{Run an add operation}
\usage{
run_add(op_node)
}
\arguments{
\item{op_node}{a greta add operation node}
}
\value{
the value of the operation
}
\description{
Run an add operation
}
