% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{run_multiply}
\alias{run_multiply}
\title{Run an multiply operation}
\usage{
run_multiply(op_node)
}
\arguments{
\item{op_node}{a greta multiply operation node}
}
\value{
the value of the operation
}
\description{
Run an multiply operation
}
