% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_functions.R
\name{run_operation}
\alias{run_operation}
\title{Run a distibution node}
\usage{
run_operation(op_node, dag)
}
\arguments{
\item{dag}{the dag the node is part of}

\item{dist_node}{a greta operation_node}
}
\value{
an altered dag containing the obtained values as a data_node
}
\description{
Run a distibution node
}
