% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_name}
\alias{same_name}
\title{Do two nodes share a name?}
\usage{
same_name(node1, node2)
}
\arguments{
\item{node1}{a greta node}

\item{node2}{a greta node}
}
\value{
logical vector of length 1
}
\description{
Do two nodes share a name?
}
