% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_functions.R
\name{sample_distribution}
\alias{sample_distribution}
\title{Sample a distibution node}
\usage{
sample_distribution(dist_node, dag)
}
\arguments{
\item{dist_node}{a greta distibution_node}

\item{dag}{the dag the node is part of}
}
\value{
an altered dag containing the simulated values as a data_node
}
\description{
Sample a distibution node
}
