% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate a from a greta model}
\usage{
simulate(m, values)
}
\arguments{
\item{m}{a greta model}

\item{values}{fixed values of some of parameters of the model}
}
\value{
TBD
}
\description{
Simulate a from a greta model
}
\examples{
TODO
}
