% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{target}
\alias{target}
\title{get the target node of a greta disitbution node}
\usage{
target(dist_node)
}
\arguments{
\item{dist_node}{a greta distribution node}
}
\value{
a greta variable node
}
\description{
get the target node of a greta disitbution node
}
