% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_field}
\alias{update_field}
\title{Update a certain 'field' of all nodes in a dag based on the prescence of
a new node}
\usage{
update_field(new_node, dag, field)
}
\arguments{
\item{new_node}{a greta node which has been construted to replace an
existing node}

\item{dag}{a greta dag}

\item{field}{a character string representing a slot in the node class}
}
\value{
the updated node_list
}
\description{
Update a certain 'field' of all nodes in a dag based on the prescence of
a new node
}
