% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_list}
\alias{which_list}
\title{a which function for lists}
\usage{
which_list(list, pred)
}
\arguments{
\item{list}{a list}

\item{pred}{a predicate: a function that takes an element of the list
and outputs a length 1 logical vector}
}
\value{
the indices of the list
}
\description{
a which function for lists
}
